# singles-collection/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
basePath: /
schemes:
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
  /single/{id}:
    x-swagger-router-controller: single
    get:
      description: Get a single
      operationId: getOne
      parameters:
        - name: id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetSingleResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      description: Update a single
      operationId: update
      parameters:
        - name: id
          description: Single id
          type: string
          in: path
          required: true
        - name: single
          description: Single properties
          in: body
          required: true
          schema:
            $ref: "#/definitions/Single"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      description: Deletes a single
      operationId: delSingle
      parameters:
        - name: id
          description: Single id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
  Single:
    type: object
    required:
      - title
      - artist
      - year
    properties:
      title:
        type: string
        description: Title of the single
      artist:
        type: string
        description: Artist of the single
      year:
        type: number
        description: Year the single was released
  GeneralResponse:
    type: object
    required:
      - success
      - description
    properties:
      success:
        type: number
        description: Returns 1 if successful
      description:
        type: string
        description: A short comment
  GetSingleResponse:
    required:
      - id
      - title
      - year
    properties:
      id:
        type: string
      title:
        type: string
      artist:
        type: string
      year:
        type: number
...
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string